requirements.txt
torch
torchvision
transformers
streamlit
opencv-python
pillow
numpy
matplotlib
grad-cam
timm
